requirements.txt
fastapi==0.68.1
google-api-core==1.30.0
google-api-python-client==2.3.0
google-auth==1.32.0
google-cloud-dialogflow==2.7.1
google-cloud-firestore==2.1.1
google-cloud-pubsub==2.4.2
requests==2.26.0
uvicorn==0.15.0
jsons==1.5.1